\name{airGP-internal}
\alias{uncollapse}
\alias{fnA}
\alias{fnB}
\alias{lengthscale.fixation}
\alias{magnitude.fixation}
\alias{logdiff}
\alias{logsum}
\alias{beta.pars}

\title{Internal airGP Function}
\description{These functions are repeatedly used by the main functions of the package "airGP". Not intended for direct use the by user.}

\keyword{internal}
